Implement workaround in setup script
The workaround script is run in the background when the wine version is
5.12 or higher. When wine bug 49532 is resolved, we need to update the
script to stop using the workaround on fixed versions. This commit also
makes winetricks run everything silently, meaning no user interaction
is required, reducing the possibility of breaking the workaround.
Finally, winetricks is run in --force mode, as it would otherwise abort
execution due to wine bug 49532 being detected.